+++ packages/docmanager/src/index.ts
@@ -1,6 +1,8 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
+
+export * from './actions';
 export * from './dialogs';
 export * from './manager';
 export * from './savehandler';
